moved UserListFilters out of UserList component
