+++ src/client/pages/Admin/UserManagement/UserManagement.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react'
 import { useAppDispatch } from '@client/store'
 import { UserManagementActions, useUsers } from '@client/store/ui/userManagement'
 import UserList from '@client/components/UserList'
-import UserListFilters from '@client/components/UserList/UserListFilters'
+import UserListFilters from '@client/components/UserListFilters'
 
 const UserManagement: React.FC = () => {
   const dispatch = useAppDispatch()
